Update how column-to-column/lit comparisons are expressed in tags
Instead of allowing an expression-like syntax inside the tag, opt for a list of values that are separated by commas. For example instead of writing "col_a >= col_b" one would write "col_a,>=,col_b". These values can essentially be thought of as a list of arguments of the gosql.Column directive. The main reasoning behind this change is to have the gosql.Column directive's tags be consistent with those of the normal "whereblock" fields which already use this syntax (using the expression-like syntax with the normal fields would be another option however that looks unnatural - a comparison operator with only one operand).