Change the supported syntax for comparison expressions in tags
This is a reversal of the design decision made in 8a09a8ca91fce9dd306be108841637df5569bbc6. The reason behind this change is to allow for multiple comparison expressions within a single JOIN in a manner that is more typical and familiar. Doing this with commas as the main separator, while possible, would not only make it more cumbersome to parse but also less comprehensible to the reader.